requirements.txt
matplotlib
numpy
scikit-image
scikit-learn
torch
tqdm
Pillow
imageio
opencv-python
ngt
GPUtill
humanize
numba